# cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/shifts": {
            "put": {
                "description": "update shift",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "update a shift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Shift",
                        "name": "shift",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Shift"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "shift"
                        }
                    }
                }
            },
            "post": {
                "description": "get shift by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "create a shift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Shift",
                        "name": "shift",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Shift"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "shift"
                        }
                    }
                }
            },
            "patch": {
                "description": "update a field shift",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "update a field shift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Shift",
                        "name": "shift",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Shift"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "shift"
                        }
                    }
                }
            }
        },
        "/shifts/{dni}": {
            "get": {
                "description": "get a shift by dni",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "get a shift by dni",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dni",
                        "name": "dni",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "shift"
                        }
                    }
                }
            }
        },
        "/shifts/{dni}/{registration}": {
            "post": {
                "description": "create a shift by dni and registration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "create a shift by dni and registration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Shift",
                        "name": "shift",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Shift"
                        }
                    },
                    {
                        "type": "string",
                        "description": "dni",
                        "name": "dni",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "registration",
                        "name": "registration",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "shift"
                        }
                    }
                }
            }
        },
        "/shifts/{id}": {
            "get": {
                "description": "get shifts by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "get a shift",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Shift id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "shifts"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes shift by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shift"
                ],
                "summary": "deletes a shift",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Shift",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "shift"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Shift": {
            "type": "object",
            "required": [
                "description",
                "dischargedate"
            ],
            "properties": {
                "dentist_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "dischargedate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "patient_id": {
                    "type": "integer"
                }
            }
        }
    }
}